+++ R/aaa.R
@@ -59,13 +59,6 @@
 )
 
 .deprecations <- list(
-  credit_roles = \(caller, param) {
-    lifecycle::deprecate_warn(
-      "0.2.0",
-      what = glue("{caller}(credit_roles)"),
-      with = I(glue("`{param} = credit_roles()`"))
-    )
-  },
   explicit_roles = \() {
     lifecycle::deprecate_warn(
       "0.2.0",